+++ TsRandomizer/LevelObjects/RoomTriggers.cs
@@ -307,9 +307,6 @@ namespace TsRandomizer.LevelObjects
 		{
 			var roomKey = new RoomItemKey(levelId, roomId);
 
-			// TODO: call this somewhere that isn't every room change
-			UpdateBestiary(level, seedOptions, gameSettings);
-
 			if (RoomTriggers.TryGetValue(roomKey, out var trigger))
 				trigger.trigger(level, itemLocations[roomKey], seedOptions, gameSettings, screenManager);
 		}
@@ -410,43 +407,6 @@ namespace TsRandomizer.LevelObjects
 			var neliste = (NPCBase)NelisteNpcType.CreateInstance(false, level, position, -1, new ObjectTileSpecification());
 
 			level.AsDynamic().RequestAddObject(neliste);
-		}
-
-		static void UpdateBestiary(Level level, SeedOptions seedOptions, SettingCollection gameSettings)
-		{
-			// TODO move text replace elsewhere
-			TimeSpinnerGame.Localizer.OverrideKey("inv_use_PlaceHolderItem1", "Nothing");
-			TimeSpinnerGame.Localizer.OverrideKey("inv_use_PlaceHolderItem1_desc", "You thought you picked something up, but it turned out to be nothing.");
-
-			var bestiary = level.GCM.Bestiary;
-			Random random = new Random(seedOptions.GetHashCode());
-			foreach (var bestiaryEntry in bestiary.BestiaryEntries)
-			{
-				if (gameSettings.ShowBestiary.Value)
-				{
-					level.GameSave.SetValue(string.Format(bestiaryEntry.Key.Replace("Enemy_", "KILL_")), 1);
-				}
-
-				foreach (var loot in bestiaryEntry.LootTable)
-				{
-					// loot.DropRate = 100;
-					if (gameSettings.LootPool.Value == "Random")
-					{
-						var item = Helper.GetAllLoot().SelectRandom(random);
-						loot.Item = item.ItemId;
-						if (item.LootType == LootType.Equipment)
-							loot.Category = (int)EInventoryCategoryType.Equipment;
-						else
-							loot.Category = (int)EInventoryCategoryType.UseItem;
-					}
-					else if (gameSettings.LootPool.Value == "Empty")
-					{
-						loot.DropRate = 0;
-						loot.Item = (int)EInventoryUseItemType.PlaceHolderItem1;
-						loot.Category = (int)EInventoryCategoryType.UseItem;
-					}
-				}
-			}
 		}
 
 		static void SpawnGlowingFloor(Level level)
